ProjectMOD7BACK/requirements.txt
django
djangorestframework
python-dotenv
requests
mysqlclient>=2.2.1
django-cors-headers
Pillow
